tests(pipelines): Split test_pipelines.py into tests/pipelines/
